rewriting tests for connected draggable
